Conflicts:

 * cmake/Findclang.cmake
 * cmake/Findlld.cmake
 * cmake/Findllvm.cmake

In master branch, more search paths were added to these files with "12"
in the path. In this commit I updated them to "13".

 * src/stage1/codegen.cpp
 * src/zig_llvm.cpp
 * src/zig_llvm.h

In master branch, ZigLLVMBuildCmpXchg is improved to add
`is_single_threaded`. However, the LLVM 13 C API has this already, and
in the llvm13 branch, ZigLLVMBuildCmpXchg is deleted in favor of the C
API. In this commit I updated stage2 to use the LLVM 13 C API rather
than depending on an improved ZigLLVMBuildCmpXchg.

Additionally, src/target.zig largestAtomicBits needed to be updated to
include the new m68k ISA.

+++ src/link/Plan9.zig
@@ -36,7 +36,7 @@ hdr: aout.ExecHdr = undefined,
 
 entry_val: ?u64 = null,
 
-got_len: u64 = 0,
+got_len: usize = 0,
 
 const Bases = struct {
     text: u64,
@@ -243,7 +243,7 @@ pub fn flushModule(self: *Plan9, comp: *Compilation) !void {
     const hdr_slice: []u8 = hdr_buf[0..hdr_size];
     var foff = hdr_size;
     iovecs[0] = .{ .iov_base = hdr_slice.ptr, .iov_len = hdr_slice.len };
-    var iovecs_i: u64 = 1;
+    var iovecs_i: usize = 1;
     var text_i: u64 = 0;
     // text
     {
